+++ packages/react-router-dom/__tests__/static-location-test.tsx
@@ -56,5 +56,31 @@ describe("A <StaticRouter>", () => {
         key: expect.any(String),
       });
     });
+
+    it("retains a non-null state when passed explicitly", () => {
+      let location!: ReturnType<typeof useLocation>;
+      function LocationChecker() {
+        location = useLocation();
+        return null;
+      }
+
+      ReactDOMServer.renderToStaticMarkup(
+        <StaticRouter
+          location={{ pathname: "/the/path", search: "?the=query", state: 0 }}
+        >
+          <Routes>
+            <Route path="/the/path" element={<LocationChecker />} />
+          </Routes>
+        </StaticRouter>
+      );
+
+      expect(location).toEqual({
+        pathname: "/the/path",
+        search: "?the=query",
+        hash: "",
+        state: 0,
+        key: expect.any(String),
+      });
+    });
   });
 });